requirements.txt
# confluent-kafka
kafka-python
python-dotenv
psycopg2-binary
flask